requirements.txt
comet_ml==3.1.12
numpy==1.18.4
opencv_python==4.5.3.56
Pillow==8.4.0
scikit_learn==1.0.1
torch==1.10.0
torchvision==0.10.0+cu111
